backend/requirements.txt
Flask==3.0.3
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
playwright==1.45.0
apscheduler==3.10.4
pyyaml==6.0.1
beautifulsoup4==4.12.3
lxml==5.2.2
Jinja2==3.1.4
cssselect
httpx[socks]
python-dotenv
loguru==0.7.2 